refactor(views): split views.py into views_process.py and views_api.py
Extract process viewer and actions into views_process.py, and JSON
API endpoints and file-serving views into views_api.py. Reduces
views.py from 2058 lines to 567.
