+++ scanning/urls.py
@@ -5,55 +5,56 @@ from django.urls import include, path
 
 from scanning.monitoring import health_check, heartbeat, sentry_fail
 from scanning.views import (
-    add_page_insert,
-    add_single_detection,
-    apply_rect_to_opinion,
-    approve_detection,
-    approve_scan,
-    assign_page,
-    bake_redactions,
-    cancel_processing,
     claim_scan,
-    compute_redactions_api,
-    delete_detection,
-    delete_page,
-    dismiss_issue,
-    # dismiss_issues,
-    export_pdf,
-    flag_issue,
-    generate_files,
     login_view,
     logout_view,
     opinion_detail,
     opinion_list,
     opinion_upload,
-    pair_opinions_api,
-    progress_api,
     queue_detail_view,
     queue_upload,
     queue_view,
-    recalculate,
-    remove_flag,
-    # repare_opinions,
-    reprocess,
-    save_margin_rect,
-    save_redaction_rect,
-    # reprocess_section_view,
     scan_detail,
     scan_list,
-    scan_process_view,
+    update_scan_status,
+)
+from scanning.views_api import (
+    add_single_detection,
+    apply_rect_to_opinion,
+    approve_detection,
+    approve_scan,
+    bake_redactions,
+    compute_redactions_api,
+    delete_detection,
+    export_pdf,
+    flag_issue,
+    generate_files,
+    pair_opinions_api,
+    remove_flag,
+    save_margin_rect,
+    save_redaction_rect,
     serve_detections,
     serve_margin_rects,
     serve_ocr_results,
     serve_opinions,
     serve_opinionscan_pdf,
-    serve_original_crop,
     serve_redacted_pdf,
     serve_redaction_rects,
+)
+from scanning.views_process import (
+    add_page_insert,
+    assign_page,
+    cancel_processing,
+    delete_page,
+    dismiss_issue,
+    progress_api,
+    recalculate,
+    reprocess,
+    scan_process_view,
+    serve_original_crop,
     serve_scan_pdf,
     start_detect,
     start_validate,
-    update_scan_status,
 )
 
 urlpatterns = [
